Fix a bug with Kamailio TLS configuration
If TLS is included in transports of AccessService, Kamailio module tls.so is added to 'layer.cfg' with config file "{{layer}}.tls.cfg". Howewer, this TLS config file is not created anywhere, so Kamailio simply fails to start.
This fix adds template for this file and makes it be created when user enables TLS transport.
A new TLS class is also added to AccessService class to make Kamailio TLS configurable by user. The most common settings are included:
- TLS version
- path to certificate
- path to private key
- custom Server Name Indication string