Allow creation of Access Domains with IP adrreses in addidition to RFC domain mames
Many SIP clients have only one 'domain' field, which implies both the hostname of SIP registrar and the 'domain' field in AOR.
A field to set SIP registrar address separately may be missing or hidden deep in the settings.
On the other hand, server with LibreSBC may not have a real domain name.
It seems to be good to allow administrator of such server create Access Domains with IP address instead of RFC domain name to simplify setup of various SIP clients.