A large rewrite of Dockerfile
1. With Docker multi-stage builds, entire process is divided for two stages: build and runtime.
In build stage all build tools and packages are installed and sources are downloaded and compiled.
In runtime stage, a smaller debian image is used, only runtime necessary packages are installed, and only binaries are copied from build stage.
This made it possible to reduce the size of container from 2.64 Gb to less than 1 Gb (!)

2. All heavy steps such as downloading tons of sources and building freeswitch and kamailio are moved to the beginning of the Dockerfile.
Now, when the changes are made only to LibreSBC code, docker uses build cache for these heavy steps, so subsequent bulds are made MUCH faster.

3. A docker volume 'libresbc' for storing Redis database and other data created by user. Now the libresbc container can be safely deleted and replaced (e.g. for update), and all user settings are persisted in real host file system (currently /var/lib/docker/volumes/libresbc)

4. WebUI is included into image

5. Built-in Redis and WebUI are enabled by default for smoother user experience. Docker best practices recommend to build ready-to-use container with all components inside it.